requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
httpx
requests
openai
tiktoken
openai-agents>=0.6.1
cohere 